requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.2
beautifulsoup4==4.12.3
readability-lxml==0.8.1
trafilatura==1.12.2
tenacity==9.0.0
python-dotenv==1.0.1

langgraph==0.2.36
langchain-core==0.3.21

openai==1.43.0
duckduckgo-search==6.3.7

pypdf==5.0.0
google-generativeai==0.8.3